Moved Swarm to Misc category
