+++ src/main/java/minegame159/meteorclient/commands/commands/swarm/SwarmFollow.java
@@ -10,7 +10,7 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import minegame159.meteorclient.commands.Command;
 import minegame159.meteorclient.commands.arguments.PlayerArgumentType;
 import minegame159.meteorclient.modules.Modules;
-import minegame159.meteorclient.modules.combat.Swarm;
+import minegame159.meteorclient.modules.misc.Swarm;
 import net.minecraft.command.CommandSource;
 import net.minecraft.entity.player.PlayerEntity;
