+++ src/main/java/minegame159/meteorclient/commands/commands/swarm/SwarmRelease.java
@@ -8,7 +8,7 @@ package minegame159.meteorclient.commands.commands.swarm;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import minegame159.meteorclient.commands.Command;
 import minegame159.meteorclient.modules.Modules;
-import minegame159.meteorclient.modules.combat.Swarm;
+import minegame159.meteorclient.modules.misc.Swarm;
 import net.minecraft.command.CommandSource;
 
 import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
