+++ src/main/java/minegame159/meteorclient/commands/commands/swarm/SwarmScatter.java
@@ -11,7 +11,7 @@ import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import minegame159.meteorclient.commands.Command;
 import minegame159.meteorclient.modules.Modules;
-import minegame159.meteorclient.modules.combat.Swarm;
+import minegame159.meteorclient.modules.misc.Swarm;
 import net.minecraft.command.CommandSource;
 
 import java.util.Random;
